Feat: 유저 인증 인가 테스트 코드 작성 및 중복 파일 제거 등
