+++ src/main/java/com/example/cacheproject/CacheProjectApplication.java
@@ -1,0 +1,13 @@
+package com.example.cacheproject;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CacheProjectApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CacheProjectApplication.class, args);
+    }
+
+}
