+++ src/main/java/com/example/cacheproject/common/dto/AuthUser.java
@@ -1,0 +1,18 @@
+package com.example.cacheproject.common.dto;
+
+import lombok.Getter;
+import com.example.cacheproject.domain.user.enums.UserRole;
+
+@Getter
+public class AuthUser {
+
+    private final Long id;
+    private final String email;
+    private final UserRole userRole;
+
+    public AuthUser(Long id, String email, UserRole userRole) {
+        this.id = id;
+        this.email = email;
+        this.userRole = userRole;
+    }
+}
