+++ src/main/java/com/example/cacheproject/domain/auth/dto/request/LoginRequestDto.java
@@ -1,0 +1,16 @@
+package com.example.cacheproject.domain.auth.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class LoginRequestDto {
+
+    @NotBlank(message = "이메일을 입력해주세요.")
+    private String email;
+
+    @NotBlank(message = "비밀번호를 입력해주세요.")
+    private String password;
+}
