+++ src/main/java/com/example/cacheproject/domain/auth/enums/TokenStatus.java
@@ -1,0 +1,6 @@
+package com.example.cacheproject.domain.auth.enums;
+
+public enum TokenStatus {
+    VALID,
+    INVALIDATED
+}
