+++ src/main/java/com/example/cacheproject/domain/auth/repository/RefreshTokenRepository.java
@@ -1,0 +1,11 @@
+package com.example.cacheproject.domain.auth.repository;
+
+import com.example.cacheproject.domain.auth.entity.RefreshToken;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+
+    Optional<RefreshToken> findByToken(String token);
+}
