+++ src/main/java/com/example/cacheproject/domain/store/Store.java
@@ -1,0 +1,32 @@
+package com.example.cacheproject.domain.store;
+
+import com.example.cacheproject.domain.user.entity.User;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor
+public class Store {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String storeName;
+
+    @Column(nullable = false, unique = true)
+    private String email;
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+
+    public Store(StoreRequestDto requestDto, User user) {
+        this.storeName = requestDto.getStoreName();
+        this.email = requestDto.getEmail();
+        this.user = user;
+    }
+}
