+++ src/main/java/com/example/cacheproject/domain/store/StoreRepository.java
@@ -1,0 +1,12 @@
+package com.example.cacheproject.domain.store;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface StoreRepository extends JpaRepository<Store, Long> {
+
+    Optional<Store> findByUserId(Long userId);
+
+    void deleteByUserId(Long userId);
+}
