+++ src/main/java/com/example/cacheproject/domain/store/StoreRequestDto.java
@@ -1,0 +1,12 @@
+package com.example.cacheproject.domain.store;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class StoreRequestDto {
+
+    private String storeName;
+    private String email;
+}
