+++ src/main/java/com/example/cacheproject/domain/store/StoreSummaryDto.java
@@ -1,0 +1,13 @@
+package com.example.cacheproject.domain.store;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+public class StoreSummaryDto {
+
+    private final String storeName;
+    private final String email;
+}
