+++ src/main/java/com/example/cacheproject/domain/user/dto/request/UserSignupRequestDto.java
@@ -1,0 +1,28 @@
+package com.example.cacheproject.domain.user.dto.request;
+
+import com.example.cacheproject.domain.store.StoreRequestDto;
+import jakarta.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class UserSignupRequestDto {
+
+    @NotBlank
+    private String email;
+
+    @NotBlank
+    private String username;
+
+    @NotBlank
+    private String password;
+
+    @NotBlank
+    private String passwordCheck;
+
+    @NotBlank
+    private String role;
+
+    private StoreRequestDto store;
+}
