+++ src/main/java/com/example/cacheproject/domain/user/dto/response/UserProfileResponseDto.java
@@ -1,0 +1,25 @@
+package com.example.cacheproject.domain.user.dto.response;
+
+import com.example.cacheproject.domain.store.Store;
+import com.example.cacheproject.domain.store.StoreSummaryDto;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import com.example.cacheproject.domain.user.entity.User;
+
+import java.time.LocalDateTime;
+
+@Getter
+@AllArgsConstructor
+public class UserProfileResponseDto {
+
+    private final Long id;
+    private final String username;
+    private final String role;
+    private final LocalDateTime createdAt;
+    private final StoreSummaryDto store;
+
+    public static UserProfileResponseDto from(User user, Store store) {
+        StoreSummaryDto storeDto = (store != null) ? new StoreSummaryDto(store.getStoreName(), store.getEmail()) : null;
+        return new UserProfileResponseDto(user.getId(), user.getUsername(), user.getUserRole().name(), user.getCreatedAt(), storeDto);
+    }
+}
