+++ src/main/java/com/example/cacheproject/domain/user/dto/response/UserUpdateResponseDto.java
@@ -1,0 +1,11 @@
+package com.example.cacheproject.domain.user.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class UserUpdateResponseDto {
+
+    private String username;
+}
