--- src/main/java/org/example/cacheproject/CacheProjectApplication.java
@@ -1,13 +1,0 @@
-package org.example.cacheproject;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CacheProjectApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CacheProjectApplication.class, args);
-    }
-
-}
