--- src/main/java/org/example/cacheproject/common/response/ErrorResponse.java
@@ -1,21 +1,0 @@
-package org.example.cacheproject.common.response;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.Getter;
-
-@Getter
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class ErrorResponse {
-
-    private final String code;
-    private final String message;
-
-    private ErrorResponse(String code, String message) {
-        this.code = code;
-        this.message = message;
-    }
-
-    public static ErrorResponse of(String code, String message) {
-        return new ErrorResponse(code, message);
-    }
-}
