--- src/main/java/org/example/cacheproject/domain/auth/dto/request/LoginRequestDto.java
@@ -1,14 +1,0 @@
-package org.example.cacheproject.domain.auth.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-
-@Getter
-public class LoginRequestDto {
-
-    @NotBlank(message = "이메일을 입력해주세요.")
-    private String email;
-
-    @NotBlank(message = "비밀번호를 입력해주세요.")
-    private String password;
-}
