--- src/main/java/org/example/cacheproject/domain/auth/enums/TokenStatus.java
@@ -1,6 +1,0 @@
-package org.example.cacheproject.domain.auth.enums;
-
-public enum TokenStatus {
-    VALID,
-    INVALIDATED
-}
