--- src/main/java/org/example/cacheproject/domain/auth/repository/RefreshTokenRepository.java
@@ -1,11 +1,0 @@
-package org.example.cacheproject.domain.auth.repository;
-
-import org.example.cacheproject.domain.auth.entity.RefreshToken;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
-
-    Optional<RefreshToken> findByToken(String token);
-}
