--- src/main/java/org/example/cacheproject/domain/user/dto/request/UserUpdateRequestDto.java
@@ -1,13 +1,0 @@
-package org.example.cacheproject.domain.user.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class UserUpdateRequestDto {
-
-    @NotBlank
-    private String username;
-}
