--- src/main/java/org/example/cacheproject/domain/user/dto/response/UserProfileResponseDto.java
@@ -1,23 +1,0 @@
-package org.example.cacheproject.domain.user.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.example.cacheproject.domain.user.entity.User;
-
-import java.time.LocalDateTime;
-
-@Getter
-@AllArgsConstructor
-public class UserProfileResponseDto {
-
-    private final Long id;
-    private final String username;
-    private final String role;
-    private final LocalDateTime createdAt;
-    private final StoreSummaryDto store;
-
-    public static UserProfileResponseDto from(User user, Store store) {
-        StoreSummaryDto storeDto = (store != null) ? new StoreSummaryDto(store.getStoreName(), store.getEmail()) : null;
-        return new UserProfileResponseDto(user.getId(), user.getUsername(), user.getRole(), user.getCreatedAt(), storeDto);
-    }
-}
