--- src/main/java/org/example/cacheproject/domain/user/dto/response/UserSignupResponseDto.java
@@ -1,13 +1,0 @@
-package org.example.cacheproject.domain.user.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class UserSignupResponseDto {
-
-    private final String message;
-    private final Long userId;
-    private final Long storeId;
-}
