--- src/main/java/org/example/cacheproject/domain/user/dto/response/UserUpdateResponseDto.java
@@ -1,11 +1,0 @@
-package org.example.cacheproject.domain.user.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class UserUpdateResponseDto {
-
-    private String username;
-}
