--- src/main/java/org/example/cacheproject/domain/user/repository/UserRepository.java
@@ -1,15 +1,0 @@
-package org.example.cacheproject.domain.user.repository;
-
-import org.example.cacheproject.domain.user.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByUsername(String username);
-    Optional<User> findByEmail(String email);
-
-    boolean existsByUsername(String username);
-    boolean existsByEmail(String email);
-}
