--- src/main/java/org/example/cacheproject/security/CustomUserDetails.java
@@ -1,53 +1,0 @@
-package org.example.cacheproject.security;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.example.cacheproject.domain.user.entity.User;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-import java.util.Collections;
-
-@Getter
-@AllArgsConstructor
-public class CustomUserDetails implements UserDetails {
-
-    private final User user;
-
-    // Spring Security 필수 메서드 구현
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return Collections.singleton(() -> "ROLE_" + user.getUserRole().name());
-    }
-
-    @Override
-    public String getPassword() {
-        return user.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return user.getUsername();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
