+++ src/test/java/com/example/cacheproject/CacheProjectApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.cacheproject;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CacheProjectApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
