--- src/test/java/org/example/cacheproject/CacheProjectApplicationTests.java
@@ -1,13 +1,0 @@
-package org.example.cacheproject;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CacheProjectApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
